backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysqlclient==2.2.0
djangorestframework-simplejwt==5.3.0
django-storages==1.14.0
python-dotenv==1.0.0
gunicorn==21.2.0
django-filter==23.2
Pillow==10.0.1
